requirements.txt
streamlit
google-generativeai
sentence_transformers
faiss-cpu
python-dotenv
tf-keras